Sprint 18 - Stage 3 - feature_subscriptions addition Checkstyle errors corrected. Link to Pull request added in Readme file. Additional corrections fixed.
