+++ ewm-service/src/main/java/ru.practicum.explorewithme/user/dto/UserWithFollowersDto.java
@@ -1,15 +1,13 @@
-package ru.practicum.explorewithme.user.dto;
+package ru.practicum.explorewithme.user.dto; //3 stage - new file
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
+import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 import java.util.List;
 
-@Getter
-@Setter
+@Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
