+++ ewm-service/src/main/java/ru.practicum.explorewithme/user/model/User.java
@@ -1,16 +1,14 @@
-package ru.practicum.explorewithme.user.model;
+package ru.practicum.explorewithme.user.model; //3 stage - add
 
 import lombok.AllArgsConstructor;
+import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
 import lombok.ToString;
 
 import javax.persistence.*;
 import java.util.List;
 
-@Getter
-@Setter
+@Data
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
@@ -20,7 +18,6 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
     private String email;
     private String name;
 
@@ -30,7 +27,6 @@ public class User {
             joinColumns = @JoinColumn(name = "follower_id"),
             inverseJoinColumns = @JoinColumn(name = "user_id"))
     private List<User> followers;
-
     @ToString.Exclude
     @ManyToMany(mappedBy = "followers")
     private List<User> followees;
